requirements.txt
PyCryptodome>= 3.10.0 